Modularise administration folder into components, views and services
