+++ frontend/src/modules/Administration/components/AdministratorDashboardViewer.js
@@ -15,22 +15,24 @@ import { Formik } from 'formik';
 import { createRef, useCallback, useEffect, useState } from 'react';
 import * as ReactIf from 'react-if';
 import * as Yup from 'yup';
-import { useSettings } from '../../../../design';
-import { SET_ERROR, useDispatch } from '../../../../globalErrors';
+import { useSettings } from 'design';
+import { SET_ERROR, useDispatch } from 'globalErrors';
+import {
+  getTrustAccount,
+  useClient
+} from 'services';
 import {
   createQuicksightDataSourceSet,
   getMonitoringDashboardId,
   getMonitoringVPCConnectionId,
   getPlatformAuthorSession,
   getPlatformReaderSession,
-  getTrustAccount,
-  updateSSMParameter,
-  useClient
-} from '../../../../services';
+  updateSSMParameter
+} from '../services';
 
 const QuickSightEmbedding = require('amazon-quicksight-embedding-sdk');
 
-const DashboardViewer = () => {
+export const DashboardViewer = () => {
   const dispatch = useDispatch();
   const client = useClient();
   const { settings } = useSettings();
@@ -449,4 +451,3 @@ const DashboardViewer = () => {
   );
 };
 
-export default DashboardViewer;
