+++ frontend/src/modules/Administration/components/index.js
@@ -1,0 +1,3 @@
+export * from './AdministrationTeams';
+export * from './AdministratorDashboardViewer';
+export * from './TeamPermissionsEditForm';
