+++ frontend/src/modules/Administration/services/getMonitoringDashboardId.js
