+++ frontend/src/modules/Administration/services/getMonitoringVPCConnectionId.js
