+++ frontend/src/modules/Administration/services/getPlatformAuthorSession.js
