+++ frontend/src/modules/Administration/services/getPlatformReaderSession.js
