+++ frontend/src/modules/Administration/services/updateTenantGroupPermissions.js
