+++ frontend/src/modules/Administration/views/AdministrationView.js
@@ -12,9 +12,8 @@ import {
 import { useState } from 'react';
 import { Helmet } from 'react-helmet-async';
 import { Link as RouterLink } from 'react-router-dom';
-import { ChevronRightIcon, useSettings } from '../../../../design';
-import AdministrationTeams from './AdministrationTeams';
-import DashboardViewer from './AdministratorDashboardViewer';
+import { ChevronRightIcon, useSettings } from 'design';
+import { AdministrationTeams, DashboardViewer } from '../components';
 
 const tabs = [
   { label: 'Teams', value: 'teams' },
