--- frontend/src/services/graphql/Tenant/getUserRoleInTenant.js
@@ -1,9 +1,0 @@
-import { gql } from 'apollo-boost';
-
-export const getUserRoleInTenant = () => ({
-  query: gql`
-    query GetUserRoleInTenant {
-      getUserRoleInTenant
-    }
-  `
-});
